added user in-out functionality

+++ server/ruby/app/graphql/types/mutation_type.rb
@@ -6,17 +6,19 @@ module Types
       field :createUser, UserType, null: false do
         argument :name, String, required: true
         argument :username, String, required: true
-        argument :bio, String, required: true
+        argument :email, String, required: true
+        argument :password, String, required: true
       end
-      def createUser(name:,username:,bio:)
-        User.create name: name, username: username, bio: bio
+      def createUser(name:,email:,username:,password:)
+        bio = "Hi everyone, my name is #{name}!  It's a pleasure to meet you!"
+        User.create name: name, email: email, username: username, password: password, bio: bio
       end
 
       field :createPost, PostType, null: false do
-        argument :userId, String, required: true
         argument :body, String, required: true
       end
-      def createPost(userId:,body:)
+      def createPost(body:)
+        userId = context[:current_user]['id']
         post = Post.create body: body
         user = User.where(id: userId).first
         userPosts = user.posts
@@ -26,11 +28,11 @@ module Types
       end
 
       field :createComment, CommentType, null: false do
-        argument :userId, String, required: true
         argument :postId, String, required: true
         argument :body, String, required: true
       end
-      def createComment(userId:,postId:,body:)
+      def createComment(postId:,body:)
+        userId = context[:current_user]['id']
         comment = Comment.create body: body
         post = Post.where(id: postId).first
         user = User.where(id: userId).first
@@ -51,6 +53,9 @@ module Types
         argument :bio, String, required: true
       end
       def updateUser(id:,name:,username:,bio:)
+        if context[:current_user]['id'].to_s != id.to_s
+          raise "Request user does not match target user"
+        end
         user = User.where(id: id).first
         user.update name: name, username: username, bio: bio
         user
@@ -62,6 +67,9 @@ module Types
       end
       def updatePost(id:,body:)
         post = Post.where(id: id).first
+        if context[:current_user]['id'].to_s != post.user.id.to_s
+          raise "Request user does not match target post author"
+        end
         post.update body: body
         post
       end
@@ -72,6 +80,9 @@ module Types
       end
       def updateComment(id:,body:)
         comment = Comment.where(id: id).first
+        if context[:current_user]['id'].to_s != comment.user.id.to_s
+          raise "Request user does not match target comment author"
+        end
         comment.update body: body
         comment
       end
@@ -81,6 +92,9 @@ module Types
         argument :id, String, required: true
       end
       def deleteUser(id:)
+        if context[:current_user]['id'].to_s != id.to_s
+          raise "Request user does not match target user"
+        end
         user = User.where(id: id).first
         user.likes.destroy_all
         user.comments.destroy_all
@@ -94,6 +108,9 @@ module Types
       end
       def deletePost(id:)
         post = Post.where(id: id).first
+        if context[:current_user]['id'].to_s != post.user.id.to_s
+          raise "Request user does not match target post author"
+        end
         post.comments.destroy_all
         post.likes.destroy_all
         Post.where(id: id).destroy_all
@@ -105,20 +122,25 @@ module Types
       end
       def deleteComment(id:)
         comment = Comment.where(id: id).first
+        if context[:current_user]['id'].to_s != comment.user.id.to_s
+          raise "Request user does not match target comment author"
+        end
         Comment.where(id: id).destroy_all
         comment
       end
 
       # Toggle Fields
       field :likeUnlikePost, LikeType, null: false do
-        argument :userId, String, required: true
         argument :postId, String, required: true
       end
-      def likeUnlikePost(userId:,postId:)
+      def likeUnlikePost(postId:)
+        userId = context[:current_user]['id']
+        puts "USER_ID_15: #{userId}"
         post = Post.where(id: postId).first
         user = User.where(id: userId).first
         postLikes = post.likes
         userLikes = user.likes
+        puts "OLD_POST_LIKES_15: #{postLikes.as_json}"
         matchingLikes = []
         for postLike in postLikes do 
           if postLike.post.id.to_s == postId && postLike.user.id.to_s == userId
@@ -126,32 +148,21 @@ module Types
           end
         end
         like = Like.create
+        puts "MATCHING_LIKES_15: #{matchingLikes.as_json}"
         if matchingLikes.length() == 0
+          puts "Add Like"
           postLikes.append(like)
           userLikes.append(like)
         else
+          puts "Remove Like"
           like = matchingLikes[0]
-          # Destroy all likes
           for matchingLike in matchingLikes do
             matchingLike.destroy
           end
-          # # Delete post likes
-          # newPostLikes = []
-          # for postLike in postLikes do 
-          #   if !postLike.post.id.to_s == postId || !postLike.user.id.to_s == userId
-          #     newPostLikes.append(postLike)
-          #   end
-          # end
-          # postLikes = newPostLikes
-          # # Delete user likes
-          # newUserLikes = []
-          # for userLike in userLikes do 
-          #   if !userLike.post.id == postId || !userLike.user.id == userId
-          #     newUserLikes.append(userLike)
-          #   end
-          # end
-          # userLikes = newUserLikes
         end
+        puts "USER__15: #{user.as_json}"
+        puts "USER_LIKES_15: #{userLikes.as_json}"
+        puts "NEW_POST_LIKES_15: #{postLikes.as_json}"
         post.update(likes: postLikes)
         user.update(likes: userLikes)
         like
